Fix project name in metadata
The project was renamed after Raku became Raku, hence the `source-url`
mentioned in the project's metadata was out of date.  Also the project's
name also seemed to be updated to use Raku rather than mention Perl6.
This commit brings the metadata up to date with the current state of the
project.