Wrap long lines to fit narrow terminals
It's still very common to use an 80 character width terminal even though
screen sizes are so much larger than they used to be.  One reason for
this is that it's easier to read as the eyes don't need to move as far
to read the text.  Another reason is that it's easier to spot changes in
diffs over shorter lines.  This commit is intended to aid readability of
the plain text for those editing the markdown in the README.